scripts/data-import-fn/requirements.txt
httpx
google-cloud-dataflow-client==0.8.6
google-cloud-storage==2.10.0
functions-framework==3.4.0